# utils/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Wupportal: An informative portal for new citizens in Wuppertal",
        "version": "1.0",
        "title": "Wupportal API",
        "contact": {
            "name": "Wupportal Team",
            "email": "info@codeschluss.de"
        }
    },
    "host": "localhost",
    "basePath": "/",
...
    "paths": {
...
        "/users": {
            "get": {
                "tags": ["user-controller"],
                "summary": "readAll",
                "operationId": "userControllerReadAll",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "sort",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "dir",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "embeddings",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "page",
                    "in": "query",
                    "required": false,
                    "type": "integer",
                    "format": "int32"
                }, {
                    "name": "size",
                    "in": "query",
                    "required": false,
                    "type": "integer",
                    "format": "int32"
                }, {
                    "name": "filter",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": ["user-controller"],
                "summary": "create",
                "operationId": "userControllerCreate",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "newUser",
                    "description": "newUser",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/UserEntity"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/resetpassword": {
            "put": {
                "tags": ["user-controller"],
                "summary": "resetPassword",
                "operationId": "userControllerResetPassword",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "username",
                    "description": "username",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/StringPrimitive"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "tags": ["user-controller"],
                "summary": "readOne",
                "operationId": "userControllerReadOne",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Resource«UserEntity»"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": ["user-controller"],
                "summary": "update",
                "operationId": "userControllerUpdate",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "newUser",
                    "description": "newUser",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/UserEntity"
                    }
                }, {
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": ["user-controller"],
                "summary": "delete",
                "operationId": "userControllerDelete",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/users/{userId}/activities": {
            "get": {
                "tags": ["user-controller"],
                "summary": "readActivities",
                "operationId": "userControllerReadActivities",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "sort",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "dir",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "embeddings",
                    "in": "query",
                    "required": false,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/{userId}/activities/{activityId}": {
            "delete": {
                "tags": ["user-controller"],
                "summary": "deleteActivity",
                "operationId": "userControllerDeleteActivity",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "activityId",
                    "in": "path",
                    "description": "activityId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/users/{userId}/organisations": {
            "get": {
                "tags": ["user-controller"],
                "summary": "readOrganisations",
                "operationId": "userControllerReadOrganisations",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": ["user-controller"],
                "summary": "addOrganisation",
                "operationId": "userControllerAddOrganisation",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }, {
                    "in": "body",
                    "name": "organisationParam",
                    "description": "organisationParam",
                    "required": true,
                    "schema": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/{userId}/organisations/{orgaId}": {
            "delete": {
                "tags": ["user-controller"],
                "summary": "deleteOrganisation",
                "operationId": "userControllerDeleteOrganisation",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "orgaId",
                    "in": "path",
                    "description": "orgaId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/users/{userId}/superuser": {
            "put": {
                "tags": ["user-controller"],
                "summary": "grantSuperuserRight",
                "operationId": "userControllerGrantSuperuserRight",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "name": "userId",
                    "in": "path",
                    "description": "userId",
                    "required": true,
                    "type": "string"
                }, {
                    "in": "body",
                    "name": "isSuperuser",
                    "description": "isSuperuser",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/BooleanPrimitive"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "BooleanPrimitive": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "boolean"
                }
            },
            "title": "BooleanPrimitive"
        },
...
        "Link": {
            "type": "object",
            "properties": {
                "deprecation": {
                    "type": "string",
                    "xml": {
                        "name": "deprecation",
                        "attribute": true,
                        "wrapped": false
                    }
                },
                "href": {
                    "type": "string",
                    "xml": {
                        "name": "href",
                        "attribute": true,
                        "wrapped": false
                    }
                },
                "hreflang": {
                    "type": "string",
                    "xml": {
                        "name": "hreflang",
                        "attribute": true,
                        "wrapped": false
                    }
                },
                "media": {
                    "type": "string",
                    "xml": {
                        "name": "media",
                        "attribute": true,
                        "wrapped": false
                    }
                },
                "rel": {
                    "type": "string",
                    "xml": {
                        "name": "rel",
                        "attribute": true,
                        "wrapped": false
                    }
                },
                "templated": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string",
                    "xml": {
                        "name": "title",
                        "attribute": true,
                        "wrapped": false
                    }
                },
                "type": {
                    "type": "string",
                    "xml": {
                        "name": "type",
                        "attribute": true,
                        "wrapped": false
                    }
                }
            },
            "title": "Link",
            "xml": {
                "name": "link",
                "namespace": "http://www.w3.org/2005/Atom",
                "attribute": false,
                "wrapped": false
            }
        },
...
        "Resource«UserEntity»": {
            "type": "object",
            "properties": {
                "_embedded": {
                    "type": "object"
                },
                "_links": {
                    "type": "array",
                    "xml": {
                        "name": "link",
                        "attribute": false,
                        "wrapped": false
                    },
                    "items": {
                        "$ref": "#/definitions/Link"
                    }
                },
                "created": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string"
                },
                "modified": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "organisationRegistrations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "phone": {
                    "type": "string"
                },
                "superuser": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "Resource«UserEntity»"
        },
...
        "StringPrimitive": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "string"
                }
            },
            "title": "StringPrimitive"
        },
...
        "UserEntity": {
            "type": "object",
            "properties": {
                "_embedded": {
                    "type": "object"
                },
                "created": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string"
                },
                "modified": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "organisationRegistrations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "superuser": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "UserEntity"
        }
    }
}